requirements.txt
aiofile==3.8.8
aiohttp==3.9.5
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
asyncio==3.4.3
attrs==23.1.0
Authlib==1.3.2
Babel==2.14.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.6.2
bs4==0.0.2
cached-property==1.5.2
cachetools==5.5.0
caio==0.9.13
certifi==2021.10.8
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
comm==0.2.2
cryptocode==0.1
cryptography==3.4.8
cssselect==1.2.0
debugpy==1.8.7
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.7
distro==1.9.0
Django==5.0.4
dnspython==2.3.0
docopt==0.6.2
docstring_parser==0.16
docutils==0.17
ebaysdk==2.1.5
email-validator==1.3.1
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.12.4
Flask==3.0.0
Flask-BabelEx==0.9.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-Sitemap==0.4.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
fqdn==1.5.1
freezegun==1.1.0
frozenlist==1.4.1
geoip2==2.9.0
gevent==21.8.0
ghostscript==0.7
google-api-core==2.22.0
google-auth==2.35.0
google-cloud-aiplatform==1.70.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==1.1.2
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
idna==2.10
infi.win32service==0.1.9
install==1.3.5
ipykernel==6.29.5
ipython==8.29.0
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jiter==0.6.1
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
Kivy==2.1.0
Kivy-Garden==0.1.5
kivymd==0.104.2
langchain==0.3.4
langchain-core==0.3.13
langchain-google-vertexai==2.0.5
langchain-ollama==0.2.0
langchain-openai==0.2.3
langchain-text-splitters==0.3.0
langsmith==0.1.137
libsass==0.20.1
lxml==4.8.0
lxml_html_clean==0.1.1
markup==0.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.7
maxminddb==2.6.1
mistune==3.0.2
msvc-runtime==14.34.31931
multidict==6.0.5
mysql-connector==2.2.9
mysql-connector-python==8.1.0
mysql-connector-python-rf==2.2.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
notebook_shim==0.2.4
num2words==0.5.10
numpy==1.26.4
ofxparse==0.21
ollama==0.3.3
openai==1.52.2
orjson==3.10.10
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pefile==2023.2.7
Pillow==10.0.1
platformdirs==3.10.0
polib==1.1.1
prometheus_client==0.21.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
psutil==5.9.0
psycopg2==2.9.5
pure_eval==0.2.3
pyasn1==0.5.0
pyasn1_modules==0.4.1
pycparser==2.21
pycryptodomex==3.18.0
pydantic==2.9.2
pydantic_core==2.23.4
pydot==1.4.2
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PyKCS11==1.5.12
pyOpenSSL==21.0.0
pyparsing==3.0.9
PyPDF2==1.26.0
pypiwin32==223
pypng==0.20220715.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
pyserial==3.5
pysmb==1.2.9.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-stdnum==1.17
pytz==2023.3
pyusb==1.2.1
pywin32==306
pywin32-ctypes==0.2.3
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.0.3
qrcode==6.1
referencing==0.35.1
regex==2024.9.11
reportlab==3.6.8
requests==2.32.3
requests-file==1.5.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rjsmin==1.1.0
rpds-py==0.20.0
rsa==4.9
schedule==1.2.1
Send2Trash==1.8.3
setuptools-odoo==3.3
setuptools-scm==8.1.0
shapely==2.0.6
sitemap-generator==0.9.13
six==1.16.0
sniffio==1.3.1
soupsieve==2.4.1
speakeasy==2.0.0
speaklater==1.3
SQLAlchemy==2.0.22
sqlite-bro==0.12.2
sqlparse==0.5.0
stack-data==0.6.3
svgwrite==1.4.3
tenacity==9.0.0
terminado==0.18.1
tiktoken==0.8.0
tinycss2==1.4.0
tornado==6.4.1
tqdm==4.65.0
traitlets==5.14.3
Tree==0.2.4
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.1
tzdata==2024.1
ujson==5.9.0
uri-template==1.3.0
urllib3==1.26.5
virtualenv==20.24.5
vobject==0.9.6.1
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.1
winpython==6.1.20230527
WTForms==3.1.1
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
yarl==1.9.4
zeep==4.1.0
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
